Refactoring continued
