+++ pkg/cmd/root/command_test.go
