--- pkg/cmd/version/options.go
@@ -1,49 +1,0 @@
-/*
-SPDX-FileCopyrightText: 2021 SAP SE or an SAP affiliate company and Gardener contributors
-
-SPDX-License-Identifier: Apache-2.0
-*/
-
-package version
-
-import (
-	"fmt"
-
-	"github.com/gardener/gardenctl-v2/internal/util"
-	"github.com/gardener/gardenctl-v2/pkg/cmd/base"
-
-	"github.com/spf13/cobra"
-)
-
-// Options is a struct to support version command
-type Options struct {
-	base.Options
-
-	// Short indicates if just the version number should be printed
-	Short bool
-	// Output defines the output format of the version information. Either 'yaml' or 'json'
-	Output string
-}
-
-// NewOptions returns initialized Options
-func NewOptions(ioStreams util.IOStreams) *Options {
-	return &Options{
-		Options: base.Options{
-			IOStreams: ioStreams,
-		},
-	}
-}
-
-// Complete adapts from the command line args to the data required.
-func (o *Options) Complete(f util.Factory, cmd *cobra.Command, args []string) error {
-	return nil
-}
-
-// Validate validates the provided options
-func (o *Options) Validate() error {
-	if o.Output != "" && o.Output != "yaml" && o.Output != "json" {
-		return fmt.Errorf(`--output must be either 'yaml' or 'json'`)
-	}
-
-	return nil
-}
